refactored utility methods
